// crm-nest-dev/src/components/notifications/enum/notifications.enum.ts
export enum notificationAction {
  MENTION = 1,
  ASSIGNED = 2,
  LEFT = 3,
  EDIT = 4,
  COMMENT_IN = 5,
  ATTACH_FILE = 6,
  ATTACH_IMAGE = 7,
  REACTION = 8,
  DUEDATE_TASK = 9
}

export enum notificationSeenAction {
  NOT_SEEN = 1,
  SEEN = 2,
}

export enum notificationSocketAction {
  CREATED = 1,
}